리팩토링 -> Action, Mutation 구조 DisplayModel로 이동, DisplayModel -> 각 RIBs로 이동
