+++ Projects/Application/Sources/ComponentNodes/Protocol/FormRegisterNodeViewable.swift
@@ -3,7 +3,7 @@ import ReactorKit
 
 protocol FormRegisterNodeViewable {
   var node: ASScrollNode { get }
-  var stateStream: Observable<RegisterDisplayModel.FormState> { get }
+  var stateStream: Observable<FormRegisterReactor.State> { get }
   var signUpButtonTapStream: Observable<Void> { get }
   var plusButtonTapStream: Observable<Void> { get }
   var keyboardDismissEventNodeTapStream: Observable<Void> { get }
