+++ Projects/Application/Sources/RIBs/Onboard/OnboardDisplayModel.swift
@@ -2,6 +2,10 @@ import UIKit
 
 enum OnboardDisplayModel {
 
+  enum Action: Equatable {
+    case postImage(UIImage)
+  }
+
   struct State: Equatable, PresentableState {
 
     static func initialState() -> State {
