+++ Projects/Application/Tests/RIBs/Authentication/Register/RegisterViewControllerSpec.swift
@@ -51,43 +51,9 @@ class RegisterViewControllerSpec: QuickSpec {
         }
 
         it("interactor의 액션 login으로 전달이 된다") {
-          expect(interactor.stub.actions.last) == RegisterPresentableAction.login
+          expect(interactor.stub.actions.last) == RegisterDisplayModel.Action.login
         }
       }
-
-//      context("사용자가 사진업로드 버튼을 입력할 경우") {
-//        beforeEach {
-//          viewController.node.registerFormNode.plusButtonNode.sendActions(forControlEvents: .touchUpInside, with: .none)
-//        }
-//
-//        context("이미지 엘범에서 사진을 선택할 경우") {
-//          beforeEach {
-//            useCase.platform.imagePickerController(
-//              UIImagePickerController(), didFinishPickingMediaWithInfo: [
-//                .originalImage: photoImage as Any,
-//                .editedImage: UIImage() as Any,
-//              ])
-//          }
-//
-//          it("사진업로드 버튼이 선택된 이미지로 변경된다") {
-//            expect(viewController.node.registerFormNode.plusButtonNode.image(for: .normal)) === photoImage
-//          }
-//
-//          it("interactor의 액션 사진으로 선택된 사진이 전달이 된다") {
-//            expect(interactor.stub.actions.last) == RegisterPresentableAction.photo(photoImage)
-//          }
-//        }
-//
-//        context("이미지 앨범을 취소한 경우") {
-//          beforeEach {
-//            useCase.platform.imagePickerControllerDidCancel(UIImagePickerController())
-//          }
-//
-//          it("interactor의 액션 사진으로 전달하지 않는다") {
-//            expect(interactor.stub.actions.last) != RegisterPresentableAction.photo(photoImage)
-//          }
-//        }
-//      }
     }
   }
 }
